requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altgraph==0.17.3
annotated-types==0.6.0
anthropic==0.37.1
anyio==3.7.1
appdirs==1.4.4
arxiv==2.1.0
astor==0.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==23.1.0
auto-py-to-exe==2.30.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.11.2
blessed==1.20.0
blinker==1.9.0
boltons @ file:///work/ci_py311/boltons_1677685195580/work
bottle==0.12.23
bottle-websocket==0.2.9
Brotli @ file:///croot/brotli-split_1714483155106/work
brotlipy==0.7.0
cachetools==5.5.0
certifi @ file:///croot/certifi_1725551672989/work/certifi
cffi @ file:///croot/cffi_1726856441404/work
cfgv==3.4.0
charset-normalizer==3.0.1
Chroma==0.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.14
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.3
conda @ file:///croot/conda_1689269889729/work
conda-content-trust @ file:///work/ci_py311/conda-content-trust_1676851327497/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1685032319139/work/src
conda-package-handling @ file:///croot/conda-package-handling_1731369017509/work
conda_package_streaming @ file:///croot/conda-package-streaming_1731366181659/work
contourpy==1.3.1
cryptography @ file:///croot/cryptography_1686613057838/work
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
docker==7.0.0
Eel==0.14.0
et-xmlfile==1.1.0
executing==2.1.0
fastapi==0.115.5
feedparser==6.0.10
filelock==3.12.4
FLAML==2.1.1
Flask==3.1.1
flatbuffers==23.5.26
fonttools==4.55.0
frozenlist==1.4.0
fsspec==2023.9.2
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
git-python==1.0.3
gitdb==4.0.10
GitPython==3.1.40
google-ai-generativelanguage==0.6.6
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.66.0
greenlet==2.0.2
grpcio==1.68.1
grpcio-status==1.62.3
h11==0.14.0
html2image==2.0.5
html2text==2024.2.26
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.28.1
huggingface-hub==0.17.3
humanfriendly==10.0
identify==2.6.3
idna @ file:///croot/idna_1714398848350/work
importlib-metadata==6.8.0
importlib-resources==6.1.0
inquirer==3.1.3
ipykernel==6.29.5
ipython==8.30.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.2
jiter==0.8.0
joblib==1.4.2
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
libmambapy @ file:///croot/mamba-split_1725563367192/work/libmambapy
litellm==1.53.1
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.6
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.1
onnxruntime==1.16.1
ooba==0.0.21
open-interpreter==0.4.3
openai==1.56.0
openpyxl==3.1.2
outcome==1.3.0.post0
overrides==7.4.0
packaging @ file:///croot/packaging_1720101850331/work
parso==0.8.4
pefile==2023.2.7
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==4.3.6
pluggy @ file:///work/ci_py311/pluggy_1676822818071/work
posthog==3.0.2
pre_commit==4.0.1
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==4.24.4
psutil==5.9.8
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pulsar-client==3.3.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyautogen==0.2.13
pycosat @ file:///croot/pycosat_1714510623388/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
pyOpenSSL @ file:///croot/pyopenssl_1678965284384/work
pyparsing==3.0.9
pyperclip==1.9.0
PyPika==0.48.9
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-editor==1.0.4
python-telegram-bot==22.1
pywin32-ctypes==0.2.0
PyYAML==6.0.1
pyzmq==26.2.0
readchar==4.0.5
referencing==0.35.1
regex==2023.10.3
requests @ file:///croot/requests_1730999120400/work
rich==13.6.0
rpds-py==0.22.0
rsa==4.9
ruamel.yaml @ file:///work/ci_py311/ruamel.yaml_1676838772170/work
selenium==4.27.1
Send2Trash==1.8.3
sgmllib3k==1.0.0
shallow-backup @ file:///home/muna/shallow-backup
shellingham==1.5.4
shortuuid==1.0.13
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
termcolor==2.3.0
tiktoken==0.7.0
tokenizers==0.14.1
tokentrim==0.1.13
toml==0.10.2
toolz @ file:///work/ci_py311/toolz_1676827522705/work
tornado==6.4.2
tqdm @ file:///croot/tqdm_1724853939799/work
traitlets==5.14.3
trio==0.24.0
trio-websocket==0.11.1
typer==0.12.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3 @ file:///croot/urllib3_1727769808118/work
uv==0.7.3
uvicorn==0.23.2
uvloop==0.18.0
vastai==0.2.5
virtualenv==20.28.0
watchfiles==0.21.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
wget==3.2
whichcraft==0.6.1
wsproto==1.2.0
yarl==1.9.2
yaspin==3.1.0
zipp==3.17.0
zope.event==4.6
zope.interface==5.5.2
zstandard @ file:///croot/zstandard_1731356346222/work
